# src/reference/api.yaml
openapi: 3.0.0
x-stoplight:
  id: 0yz6u3krgrvuh
info:
  title: openapi.yml
  version: '1.0'
servers:
  - url: 'http://localhost:3000'
paths:
  /:
    get:
      summary: Root
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    x-stoplight:
                      id: shbvpjwk98cjd
                required:
                  - message
      operationId: get-root
    parameters: []
  /protected:
    get:
      summary: Your GET endpoint
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    x-stoplight:
                      id: ta1llbhy0q0qi
      operationId: get-protected
      security:
        - JWT: []
  /auth/password:
    post:
      summary: Login with email and password
      operationId: post-auth-password
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    x-stoplight:
                      id: ue02k8h47nhl0
                required:
                  - token
        '400':
          $ref: '#/components/responses/ErrorResponse'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  x-stoplight:
                    id: ftt3dnrlc9pii
                password:
                  type: string
                  x-stoplight:
                    id: mjxnzdl7muryl
              required:
                - email
                - password
  /auth/verify:
    post:
      summary: Login with email and password
      operationId: post-auth-verify
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    x-stoplight:
                      id: ue02k8h47nhl0
                  message:
                    type: string
                    x-stoplight:
                      id: dfpcznwjaqoze
                required:
                  - token
      parameters:
        - schema:
            type: string
          in: cookie
          name: auth
          required: true
    parameters: []
  /auth/google:
    post:
      summary: Login with email and password
      operationId: post-auth-google
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    x-stoplight:
                      id: ue02k8h47nhl0
                required:
                  - token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                accessToken:
                  type: string
                  x-stoplight:
                    id: ritm50grjl358
              required:
                - accessToken
    parameters: []
  /refresh:
    get:
      summary: Your GET endpoint
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    x-stoplight:
                      id: 0nqeokosfo1hm
      operationId: get-refresh
      security: []
  /register:
    post:
      summary: Register a new user
      operationId: post-register
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    x-stoplight:
                      id: 2t12pk6197xgg
                  message:
                    type: string
                    x-stoplight:
                      id: ckyul3n07lvi3
                required:
                  - token
                  - message
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
      description: ''
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  x-stoplight:
                    id: 09c81lj7x8pxk
                password:
                  type: string
                  x-stoplight:
                    id: c6qb90gbmy0va
              required:
                - email
                - password
  /games:
    get:
      summary: All games
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                x-examples:
                  Example 1:
                    - name: string
                      description: string
                      date: 0
                      image: string
                      vote: 0
                type: array
                items:
                  $ref: '#/components/schemas/GameBase'
              examples: {}
      operationId: get-games
      description: ''
      security:
        - JWT: []
    parameters: []
  '/games/{id}':
    parameters:
      - schema:
          type: string
        name: id
        in: path
        required: true
    get:
      summary: Your GET endpoint
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GameBase'
                  - type: object
                    x-stoplight:
                      id: 1kfqf9dhakzt9
                    properties:
                      fullImage:
                        type: string
                        x-stoplight:
                          id: aiih7qovzd0nc
                      longDescription:
                        type: string
                        x-stoplight:
                          id: p5k3euqodeh1q
                      tags:
                        type: array
                        x-stoplight:
                          id: vrwo6qlfbxsnj
                        items:
                          x-stoplight:
                            id: bxywopqqjca1l
                          type: string
                      publisher:
                        type: string
                        x-stoplight:
                          id: hfkvkhzg1t7r8
                    required:
                      - fullImage
                      - publisher
        '404':
          $ref: '#/components/responses/ErrorResponse'
      operationId: get-games-id
      security:
        - JWT: []
  /users/me:
    delete:
      summary: ''
      operationId: delete-users-me
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    x-stoplight:
                      id: orgrr6pgvnc29
      security:
        - JWT: []
  '/games/{id}/votes':
    parameters:
      - schema:
          type: string
        name: id
        in: path
        required: true
    post:
      summary: ''
      operationId: post-games-id-votes
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  vote:
                    $ref: '#/components/schemas/Vote'
                required:
                  - vote
        '404':
          $ref: '#/components/responses/ErrorResponse'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                vote:
                  $ref: '#/components/schemas/Vote'
              required:
                - vote
        description: ''
    get:
      summary: ''
      operationId: get-games-id-votes
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  vote:
                    $ref: '#/components/schemas/Vote'
        '404':
          $ref: '#/components/responses/ErrorResponse'
      security:
        - JWT: []
components:
  schemas:
    GameBase:
      type: object
      x-examples:
        Example 1:
          name: string
          description: string
          year: 0
          image: string
          vote: 0
      properties:
        id:
          type: string
          x-stoplight:
            id: 1yocliz03o58b
        name:
          type: string
        description:
          type: string
        year:
          type: integer
        image:
          type: string
        vote:
          $ref: '#/components/schemas/Vote'
      required:
        - id
        - name
        - description
        - year
        - image
    Vote:
      title: Vote
      x-stoplight:
        id: ufbd9u7lcpvek
      type: number
      minimum: 1
      maximum: 5
      format: float
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
  responses:
    ErrorResponse:
      description: Example response
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                x-stoplight:
                  id: 1pivfjtt3f5nw